feat: latest css styles
