+++ react/components/CustomGridItemBig.tsx
@@ -1,24 +1,14 @@
-import React, { ReactNode } from 'react'
-import { useCssHandles } from 'vtex.css-handles'
-
-import './styles.css'
-
+import React, { ReactNode } from 'react';
+import styles from './styles.css'
 
 type Props = {
   element: ReactNode
 }
 
 const CustomGridItemBig = ({ element }: Props) => {
-
-  const CSS_HANDLES = [
-    "custom-grid__item-big",
-  ]
-  const handles = useCssHandles(CSS_HANDLES)
-
-  console.log("mi elemento es:", element)
-  return (
-    <div className={handles['custom-grid__item-big']}>{element}</div>
-  )
+  return <div className={styles.grid__itemBig}>
+    {element}
+  </div>
 }
 
 export default CustomGridItemBig
