+++ react/components/styles.css
@@ -1,4 +1,59 @@
-.custom-grid__item-big {
-  width: calc(50% -20px);
-  background-color: deeppink;
+.grid__1,
+.grid__2,
+.grid__3,
+.grid__4 {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: center;
+  padding: 20px;
+}
+
+.grid__2>div:nth-child(1) {
+  order: 2;
+}
+
+.grid__2>div:nth-child(2) {
+  order: 1;
+}
+
+.grid__2>div:nth-child(3) {
+  order: 3;
+}
+
+.grid__3>div:nth-child(1) {
+  order: 1;
+}
+
+.grid__3>div:nth-child(2) {
+  order: 3;
+}
+
+.grid__3>div:nth-child(3) {
+  order: 2;
+}
+
+.grid__4>div:nth-child(1) {
+  order: 3;
+}
+
+.grid__4>div:nth-child(2) {
+  order: 3;
+}
+
+.grid__4>div:nth-child(3) {
+  order: 1;
+}
+
+.grid__itemBig {
+  width: 40%;
+  margin: 10px;
+  height: auto;
+  max-height: 500px;
+  display: flex;
+}
+.grid__itemSmall {
+  width: 25%;
+  margin: 10px;
+  height: auto;
+  max-height: 500px;
 }
